Expanded Movedex
The movedex should now contain all available moves.

Changed the contact flag in the move/9 predicate with a list of given
flags. This change had little effect on the battle processor and the
saved teams.

Added move_has_flag/2 predicate to check if a certain move has a
specific flag or not.

Expanded todo list.