Reworded 'Move' to 'Action'
'Move' was used to refer to either a pokemon's move or the player's move
choice in battle (being an active pokemon's move or a switch to a
different pokemon).
This is now altered and the version refering to both a pokemon's move
and a switch is now called an action.

This does nothing but help to distinguish the two better.