New Predicate for Primary Status Conditions
There is a new predicate to access the pokemon's primary status
condition's category.

This can be used to see if a pokemon is poisoned or not when it is not
relevant to distinguish between poison or toxin or to see how long the
toxin already has been active.

It also only shows if a pokemon is asleep or not, without showing how
many turns it will remain sleeping.