Increased Difficulty
Rot now builds up a search tree of depth 2, thus increasing the difficulty to win against him (at least in theory).

This is a little experimental as it occured to run out of stack while testing the battle.

Further more a singleton variable in the battle_interface was replaced by _