(Eventually) Fixed Out Of Local Stack Error
Added two cuts to the source code. Both are intended to cut away choice points created by prolog to which the program is not supposed to backtrack to.

It seems that this cuts help to prevent Out Of Local Stack errors as the stack is no longer flooded with unnecessary choice points.

Further testing needed, but it looks promising