Altered Stab Predicate
stab/3 is now stab/2 and has altered arguments. This change was made to increase the difference between stab/3 and calculate_stab/3.

stab/2 now has pokemon data as it's first and the move type as it's second argument. It is true if the given pokemon has STAB for a move of the given type. It no longer returns a factor indicating the STAB value.