Added Effectiveness Predicates
Added two new predicates  and  indicating if a given pokemon gets effectively hit by a given move type (or not very effectively, respectively).

If one predicate is false the other one is by no means automatically true, as a pokemon has the possibility to be hit neutrally which is neither *very effective* nor *not very effective*.

If one predicate is true the other can not also be true. If both are false the move type hits the pokemon neutrally.