First Step for Expert System
Added the first part of a system intended to figure out the opponent's (being the player in the eyes of Rot) team setup including move sets and ev/dv distributions of the pokemon

At current state it is possible to initialize with rot_init/1 commiting a list of pokemon names representing the team pokemon of the player. Rot asserts (very basic and unspecific) knowledge about those pokemon. By calling derive_team/1 one can derive a specific version of the team data, as the data itself is asserted with possible domains for the status values in the first place. This derivations are intended to use for the search tree creation rot is using to determine his next action, making it more fair to the player, as rot does not know everything about the player's team from the very begining of the battle, thus rendering closer to real battles.

In the future those data will be updated after every turn, drawing closer and closer to the true values and thus representing (kind of) a humans observation of the opponents team and learning what pokemon has what role in said team.