Overhauled Message System
The message system is now designed in such a way, that the underlying data structure is of no interest for the user. Previously it was working on a stack and had predicates making it possible to take advantage of it, like the push_messages/3 predicate, where it was assumed that it recieved stacks (anti-chronological message lists) as input and could be abused to push two stacks together without having one of them created by the message system in the first place.
Code like push_messages([b,a], [d,c], Msg) was possible, resulting in Msg = [d,c,b,a] (let a,b,c,d be the chronological order of those messages), making it unnecessarily hard for someone wanting to e.g. alter the message system to work with queues, as the sub lists [d,c] and [b,a] are anti-chronological and the code then would produce a result Msg = [b,a,d,c], thus not having any chronological semantic at all.

Now the message system works with add_messages/3 that creates something called a message collection. The added masages (1st arg) are a list of chronological order to be added to a message collection (2nd arg), resulting in an new message collection (3rd arg).
The collections internal representation is no longer set in stone, but one can now easily alter the way messages are handled internally without having to change all the code where messages are used, but only by editing `tools/message_handling.pl`.

For convenience one can always use [] as an empty collection. To create a new collection just add messages to the empty collection.

This design choice to let the message system handle itself and just provide an API for it should have been made from the start.