Make Use of New Message Frame
The messages now have a shorter style to make use of the new message frame. As the frame corresponds to a given player and holds information about the two active pokemon the messages no longer need to carry this information themselves, but can be assumed to relate to the aforementioned player's active pokemon.
If a message is meant to relate to the opposing player's active pokemon the message can be wrapped in a target/1 predicate what correctly switches the logic for the message to relate to the target pokemon.

As some message setups now are redundant they got removed:
  - move_use_message/3
  - move_critical_message/2
  - move_effectiveness_message/2

The effectiveness and critical tags of the calculate_damage/5 predicate now are predicates themselves. See the PlDoc for more information, but this tags now can be used directly as messages.