Added First Step to Message Evaluation
Added predicates allowing Rot to evaluate the message frames of the battle.

Currently Rot only updates the move set data of the player's known pokemon data and does not use it any further. But then again, this is only the first step.

The user can observe that Rot updates the data by using the newly added known(Pokemon) and derived(Pokemon) user commands.