Removed Attack Stat Evaluation Predicate
On writing the evaluation of defense and hit point stats I noticed that
it could be better to merge the predicates for stat evaluation into
their corresponding move evaluation predicate clauses.

Offensive stats are now evaluated in the clause of rot_evaluate_damage/6
that handles moves used by the player.