Enhanced Move Initialization
Instead of assuming that each pokemon only knows tackle (which does not
make sense since there are pokemon not learning tackle at all), the
initialization works roughly as follows:

First for each type the pokemon has it is assumed that it knows a
offensive move of that type.

The remaining move slots get filled by moves of types increasing the
type coverage, aiming to hit very effectively as many pokemon as
possible.

For each type the strongest to the pokemon available move is choosen.

The algorithm for this lacks to add utility moves, like ones with higher
priority or any status move at all an assumes for now, that every
pokemon is played offensively and offensively only.