Updated Movedex and Added Bad Poison
The movedex had some little flaws (some effect probabilities were set to 0%, some status value changes were not correctly targeted, bad poison was saved as poison with duration between(15,15), etc).

Also bad poison (formerly known as toxin, which was the german phrase for it...) could not be handled correctly, as effects (like toxic) causing it were saved under the wrong format.

Now it is called bad poison, flagged as bad-poison and will be applied as poison(Counter) to the pokemon data. It will not be bad-poison(Counter), as -/2 is a predicate in prolog and bad-poison(C) =.. [Functor|Args] will result in Functor = -
If somewhere in the code =.. is used to extract the functor of an unknown primary status condition this could lead to unpleasant results.