Updated Pokedex
The pokedex is now a directory by itself, containing not only data about what pokemon has which stats or abilities available, but also more information about the pokemon forms.

There is access to whether a certain pokemon form is a battle only form or not, and there is a predicate to map a certain form to a species, what makes it easier to check if a pokemon of a certain species already is in the team.

These changes were made with the team builder in mind, thus they allow easy team validation.
Teams with e.g. two different Rotom forms or an Aegislash already in it's blade form can now be detected easier.

Furthermore the abilities now reflect the style choice already made for moves, that instead an ability being saved as 'solar power' it is saved as solar-power.