Rectified Switching out of Fainted Pokemon
When a pokemon had fainted before it could use it's move, the player was prompted to switch it out. This happened before the calculation of the end of turn (including damage by burn or poison)

Now the move useage will just be skipped and the correspinding player will be prompted to switch the fainted pokemon out after the calculation of his end of turn.