Security Clause for Damage Evaluation
rot_evaluate_damage/8 tended to fail, if the formula used could not properly be reconstructed by clpfd due to rounding errors. This resulted in infinite loops.

The security clause is there to just skip the evaluation if the calculation fails. In this way the known data of the player's pokemon remain unaltered, but at least neither wrong nor does the game enter an infinite loop.