Fixed End of Turn Evaluation
Due to the recent update from the master branch and its overhaul of the fainted pokemon handling, there were some issues with the end of turn evaluation, as it used variables no longer needed, thus setting the game state assumed by Rot to a partial unbound one. This should be fixed.

Furthermore the aforementioned update also included a polished mechanic for bad poisoning, what was not a part of the evaluation at all. Now Rot should evaluate bad poisoning correctly.