Fixed clpfd Division
As of SWI-Prolog 7.2.3, due to the update to the clpfd library, (/)/2 is no longer supported to use in constrains. As the evaluation of the damage formula relied on this division, the system did no longer work properly but failed as soon as an offensive move had to be evaluated.

Now (//)/2 is used and the system should work properly again, even under SWI-Prolog 7.2.3