Rearranged Rot's Code
Now all predicates asserting data as rot/1 predicates are located either in rot/expert/init.pl or rot/expert/data_access.pl

The old API was not changed, but serves now as an additional layer of abstraction covering the assert and retract calls. This should make it easier to alter the intern data structure without any impact on how the code runs.