Added Save States
Added predicates to save the current state of the game to a save file. The game can then be started again from the exact same point by calling test_save_state.

In this iteration every turn the state will be saved. This is for debug reasons and should be changed in the future.