Added Instances to Rot
Instead of one instance, Rot is now capable of asserting and handling data for different instances that are asserted at the same time to the memory, but may refer to totally different battles.

This can be used in the future to let Rot battle itself, where each player is an individual instance of Rot.