(Maybe) Fixed Rot's Hit Point Guessing
As Rot tries to evaluate the status values of its opponent, it also evaluates the hit points of the opponent. Apparently there was a bug resulting in the evaluated hit points being higher than the actual hit points of the opponent's pokemon (often 1 to 10 points higher).

This resulted in system crashes whenever the evaluated hit points where evaluated high enough to make a difference in the possible damage percentages. (Hhit points are evaluated by the possible damage range of a move and the percentage of hit points the target has lost).

As from now on the percentages themselves are calculated by clpfd with a variance of 0..1 (to account rounding errors).

This seems to work stable now.