Rot Can now Battle Itself
By calling test_rot_battle/0 or test_rot_battle_small/0 Rot engages in a battle against another Rot-instance.

In the future it will be possible to assign each instance its own search algorithm and rating heuristic to allow for better comparison.

Sometimes the move evaluation fails, resulting in infinite loops, but this is a problem of the evaluation itself and may be fixed in the future.