Fixed Hit Point Evaluation Fix.
It seems that, instead of fixing the hit point evaluation, it just stopped working. Classic!

I altered the formula to not solve for the new maximum hit points but the damage instead, as clpfd is still able to recalculate domains on either side of (#=)/2.
This has the side effect, that the formula stays exactly the same as used to determine the possible damage range and should be less error-prone to rounding errors caused by integer division (at least I could imagine something like that).

I let Rot battle itself 15 times, and in those 15 battles no system failure could be observed. But to quote Dijkstra: "Program testing can be a very effective way to show the presence of bugs, but is hopelessly inadequate for showing their absence."