Baked Ignoring of Message Evaluation into API
The predicates in Rot's API responsible for transmitting the message frames to evaluate on now are responsible for ignoring this transmission if rot(searching) is true, e.g. Rot is in it's search algoithm.

Before this was handled by the evaluation system itself, but now it feels much more intuitive to me.