Fixed EV Labeling
Before only 3 of the 6 status value's EV got labeled, as - assuming all 510 EV are spent - if two stats are on 252 EV and the third on the remaining 6, the others should default to 0.

Though this was right it was not accounting the possibility of not fully spent EV, like all being 0. This opened up the possibility of labeling 3/6 values being not sufficient, rendering the remaining values as unlabeled.

This caused problems as these unlabeled values still were unbound variables.