Added cut_tree/4
This is a handy little predicate to cut away some of the search trees first subtrees. Its intention is to enable Rot to search for an action to choose and also give input to how Rot's chosen action has to look like.

For example one can cut away all subtrees created by Rot's possible actions that do not unify with switch(_) to force Rot to find a switch(_) action.