Added Minmax Prediction Search
The default minmax assumes that the player knows what Rot's move will be and expects the player to counter it. So the found actions are (Counter_rot_action, Rot_action).

The idea of minmax prediction search is to counter the counter, resulting in (Player_action, Counter_action).

If the algorithm is superior in its current implementation is still in testing.