Added a Way to Search at Forced Switches
Currently and prior to this commit, forced switches were handled at
random choice, by Rot. That was bad.

I implemented a way, to create a tree, that has its first available
actions predestined. However, this ended up in infinite loops. That was
bad, too.

The code is implemented, but currently not active. This is a task for my
future self, to implement it correctly.