Revert "README is no longer an .md-File"
This reverts commit 51d9135979c32c9680c25c529faeebc37774c73f.

This was not my brightest idea. I thought GitLab only cares for files named README, independent on the file extension, but I was wrong. I removed it, as I thought README without the .md looks cleaner to hand out on CD with my thesis.

Now I hate this change and revert it.